+++ lib/posthog_flutter.dart
@@ -2,6 +2,7 @@ library posthog_flutter;
 
 export 'src/posthog.dart';
 export 'src/posthog_config.dart';
+export 'src/posthog_event.dart';
 export 'src/posthog_observer.dart';
 export 'src/posthog_widget.dart';
 export 'src/replay/mask/posthog_mask_widget.dart';